Combine consumers/batch consumers
