+++ lib/deimos.rb
@@ -9,9 +9,7 @@ require 'deimos/avro_data_decoder'
 require 'deimos/producer'
 require 'deimos/active_record_producer'
 require 'deimos/active_record_consumer'
-require 'deimos/active_record_batch_consumer'
 require 'deimos/consumer'
-require 'deimos/batch_consumer'
 require 'deimos/configuration'
 require 'deimos/instrumentation'
 require 'deimos/utils/lag_reporter'
@@ -133,17 +131,39 @@ module Deimos
         delivery = listener.delivery
 
         # Validate that Deimos consumers use proper delivery configs
-        if handler_class < Deimos::BatchConsumer
-          unless delivery == 'inline_batch'
-            raise "BatchConsumer #{listener.handler} must have delivery set to"\
-              ' `inline_batch`'
+        # if handler_class < Deimos::BatchConsumer
+        #   unless delivery == 'inline_batch'
+        #     raise "BatchConsumer #{listener.handler} must have delivery set to"\
+        #       ' `inline_batch`'
+        #   end
+        # elsif handler_class < Deimos::Consumer
+        #   if delivery.present? && !%w(message batch).include?(delivery)
+        #     raise "Non-batch Consumer #{listener.handler} must have delivery"\
+        #       ' set to `message` or `batch`'
+        #   end
+        # end
+
+        next unless handler_class < Deimos::Consumer
+
+        if delivery == 'inline_batch'
+          if handler_class.instance_method(:consume_batch).owner != handler_class
+            raise "BatchConsumer #{listener.handler} does not implement `consume_batch`"
           end
-        elsif handler_class < Deimos::Consumer
-          if delivery.present? && !%w(message batch).include?(delivery)
-            raise "Non-batch Consumer #{listener.handler} must have delivery"\
-              ' set to `message` or `batch`'
-          end
+        elsif handler_class.instance_method(:consume).owner != handler_class
+          raise "Non-batch Consumer #{listener.handler} does not implement `consume`"
         end
+
+        # if handler_class.instance_method(:consume_batch).owner == handler_class
+        #   unless delivery == 'inline_batch'
+        #     raise "BatchConsumer #{listener.handler} must have delivery set to"\
+        #       ' `inline_batch`'
+        #   end
+        # elsif handler_class.instance_method(:consume).owner == handler_class
+        #   if delivery.present? && !%w(message batch).include?(delivery)
+        #     raise "Non-batch Consumer #{listener.handler} must have delivery"\
+        #       ' set to `message` or `batch`'
+        #   end
+        # end
       end
     end
   end
