+++ spec/batch_consumer_spec.rb
@@ -1,14 +1,12 @@
 # frozen_string_literal: true
 
-require 'deimos/batch_consumer'
-
 # :nodoc:
 module ConsumerTest
-  describe Deimos::BatchConsumer do
+  describe Deimos::Consumer do
 
     prepend_before(:each) do
       # :nodoc:
-      consumer_class = Class.new(Deimos::BatchConsumer) do
+      consumer_class = Class.new(Deimos::Consumer) do
         schema 'MySchema'
         namespace 'com.my-namespace'
         key_config field: 'test_id'
@@ -104,7 +102,7 @@ module ConsumerTest
       end
 
       it 'should decode plain keys for all messages in the batch' do
-        consumer_class = Class.new(Deimos::BatchConsumer) do
+        consumer_class = Class.new(Deimos::Consumer) do
           schema 'MySchema'
           namespace 'com.my-namespace'
           key_config plain: true
@@ -121,7 +119,7 @@ module ConsumerTest
     describe 'timestamps' do
       before(:each) do
         # :nodoc:
-        consumer_class = Class.new(Deimos::BatchConsumer) do
+        consumer_class = Class.new(Deimos::Consumer) do
           schema 'MySchemaWithDateTimes'
           namespace 'com.my-namespace'
           key_config plain: true
@@ -204,7 +202,7 @@ module ConsumerTest
     describe 'logging' do
       before(:each) do
         # :nodoc:
-        consumer_class = Class.new(Deimos::BatchConsumer) do
+        consumer_class = Class.new(Deimos::Consumer) do
           schema 'MySchemaWithUniqueId'
           namespace 'com.my-namespace'
           key_config plain: true
