+++ spec/handlers/my_consumer.rb
@@ -1,5 +1,11 @@
 # frozen_string_literal: true
 
 module ConsumerTest
-  class MyConsumer < Deimos::Consumer; end
+  # :no-doc:
+  class MyConsumer < Deimos::Consumer
+    # :no-doc:
+    def consume(_payload, _metadata)
+
+    end
+  end
 end
